move invoices version controlling to database

+++ BuyAndSell/Business/BuyInvoiceBusiness.cs
@@ -76,20 +76,18 @@ namespace BuyAndSell.Business
             catch { }
             return 0;
         }
-        public override bool EditInvoice(int lastNumber, DataTable invoicetable, DataTable invoiceitems)
+        public override DatabaseSaveResult EditInvoice(int lastNumber, DataTable invoicetable, DataTable invoiceitems)
         {
             var connection = _database.GetConnection();
             var transaction = _database.BeginTransaction(connection);
-            if (Framework.Utilities.ArrayComparator.AreEqual(GetInvoiceVersion(lastNumber, Invoice.InvoiceType.Buying, connection, transaction), invoicetable.Rows[0].Field<byte[]>(Invoice.VersionColumnName))
-                && RemoveInvoice(lastNumber, connection, transaction) && Save(invoicetable, invoiceitems, connection, transaction))
-            {
+            var res = RemoveInvoiceItems(lastNumber, connection, transaction) ? DatabaseSaveResult.Saved : DatabaseSaveResult.Error;
+            if (res == DatabaseSaveResult.Saved) res = Save(invoicetable, invoiceitems, connection, transaction);
+            if (res == DatabaseSaveResult.Saved)
                 _database.CommitTransaction(transaction);
-                connection.Close();
-                return true;
-            }
-            _database.RollbackTransaction(transaction);
+            else
+                _database.RollbackTransaction(transaction);
             connection.Close();
-            return false;
+            return res;
         }
 
         public override bool RemoveInvoice(int number)
@@ -111,14 +109,9 @@ namespace BuyAndSell.Business
         {
             try
             {
-                var invoice = FullLoadBuyInvoice(number, connection, transaction);
-                foreach (BuyInvoiceItem x in invoice.InvoiceItems)
-                {
-                    var q = ItemQuantity(x.ItemRef, x.StockRef, connection, transaction);
-                    q.Quantity -= x.Quantity;
-                    _database.Save(q, connection, transaction);
-                    _database.Delete(x, connection, transaction);
-                }
+                var invoice = GetInvoiceModel(number, connection, transaction);
+                if (!RemoveInvoiceItems(number, connection, transaction))
+                    return false;
                 _database.Delete(invoice, connection, transaction);
             }
             catch
@@ -128,28 +121,45 @@ namespace BuyAndSell.Business
 
             return true;
         }
-        public override bool SaveInvoice(DataTable invoicetable, DataTable invoiceitems)
+        private bool RemoveInvoiceItems(int number, SqlConnection connection, SqlTransaction transaction)
+        {
+            try
+            {
+                var invoice = FullLoadBuyInvoice(number, connection, transaction);
+                foreach (BuyInvoiceItem x in invoice.InvoiceItems)
+                {
+                    var q = ItemQuantity(x.ItemRef, x.StockRef, connection, transaction);
+                    q.Quantity -= x.Quantity;
+                    _database.Save(q, connection, transaction);
+                    _database.Delete(x, connection, transaction);
+                }
+            }
+            catch
+            {
+                return false;
+            }
+
+            return true;
+        }
+        public override DatabaseSaveResult SaveInvoice(DataTable invoicetable, DataTable invoiceitems)
         {
             var connection = _database.GetConnection();
             var transaction = _database.BeginTransaction(connection);
 
-            if (Save(invoicetable, invoiceitems, connection, transaction))
-            {
+            var res = Save(invoicetable, invoiceitems, connection, transaction);
+            if (res != DatabaseSaveResult.Saved)
+                _database.RollbackTransaction(transaction);
+            else
                 _database.CommitTransaction(transaction);
-                connection.Close();
-                return true;
-            }
 
-            _database.RollbackTransaction(transaction);
             connection.Close();
-            return false;
+            return res;
         }
-        private bool Save(DataTable invoicetable, DataTable invoiceitems, SqlConnection connection, SqlTransaction transaction)
+        private DatabaseSaveResult Save(DataTable invoicetable, DataTable invoiceitems, SqlConnection connection, SqlTransaction transaction)
         {
             BuyInvoice invoice = new BuyInvoice();
             invoice.MapToModel(invoicetable.Rows[0]);
-            invoice.Id = -1;
-
+            
             var items = new List<BuyInvoiceItem>();
             foreach (DataRow row in invoiceitems.Rows)
             {
@@ -163,25 +173,26 @@ namespace BuyAndSell.Business
             return Save(invoice, connection, transaction);
         }
 
-        private bool Save(BuyInvoice invoice, SqlConnection connection, SqlTransaction transaction)
+        private DatabaseSaveResult Save(BuyInvoice invoice, SqlConnection connection, SqlTransaction transaction)
         {
             try
             {
                 if (invoice.UserRef < 1 || invoice.PartyRef < 1)
-                    return false;
+                    return DatabaseSaveResult.Error;
 
                 if (invoice.Number == -1)
                     invoice.Number = GetLastInvoiceNumber(connection, transaction) + 1;
 
-                _database.Save(invoice, connection, transaction);
+                if (_database.Save(invoice, connection, transaction) == DatabaseSaveResult.AlreadyChanged)
+                    return DatabaseSaveResult.AlreadyChanged;
 
                 if (_database.GetAllDataset<BuyInvoice>(connection, transaction, "Number=" + invoice.Number, null, 2).Tables[0].Rows.Count != 1)
-                    return false;
+                    return DatabaseSaveResult.Error;
 
                 foreach (BuyInvoiceItem x in invoice.InvoiceItems)
                 {
                     if (x.ItemRef < 1 || x.StockRef < 1)
-                        return false;
+                        return DatabaseSaveResult.Error;
                     x.InvoiceRef = invoice.Id;
                     _database.Save(x, connection, transaction);
 
@@ -197,9 +208,9 @@ namespace BuyAndSell.Business
             }
             catch
             {
-                return false;
+                return DatabaseSaveResult.Error;
             }
-            return true;
+            return DatabaseSaveResult.Saved;
         }
 
         public override bool IsInvoiceNumberValid(int num)
